pyproject.toml
[project]
name = "rest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "flasgger>=0.9.7.1",
    "flask>=3.1.2",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
]
